P4:24529

+++ bld/vi/h/rtns2.h
@@ -34,7 +34,7 @@
 void PositionScrollThumb( void );
 
 /* biosos2.c */
-void MyVioShowBuf( unsigned, int );
+void MyVioShowBuf( unsigned short, unsigned short );
 
 /* biosqnx.c */
 int KeyboardInit( void );
